** WIP ** [NTOS:KD] Implement Win8.1+ KdAcquire/ReleaseDebuggerLock()
KdAcquireDebuggerLock() elevates the IRQL to DISPATCH_LEVEL before
acquiring the debugger spinlock. KdReleaseDebuggerLock() does the
reverse.

*******************************
** ReactOS-specific addition **
**

Because the debugger spinlock is **ALSO** acquired by KdPollBreakIn(),
that can be (and is) invoked by interrupt handlers (the clock handler,
via KeUpdateSystemTime()), this spinlock **MUST** be acquired after
disabling interrupts. (And getting released before restoring them.)

This addition is ReactOS-specific, because our implementation of
KeTryToAcquireSpinLockAtDpcLevel() will bugcheck if the caller has
already acquired the spinlock. (Windows' implementation doesn't.)

Indeed, suppose there is driver code that acquires the debugger spinlock
(for example, what KdDisable/EnableDebuggerWithLock(TRUE) does, when
invoking KdEnable/DisableDebugger() from a driver...). The spinlock is
acquired at DISPATCH_LEVEL.
This code can be interrupted by the clock interrupt that periodically
invokes KdPollBreakIn(), which in turns tries to acquire the spinlock.
The interrupt runs within the thread's execution context, hence it is
the same thread that tries to acquire the spinlock again.
Of course this is "known" by KdPollBreakIn() and is why it uses
KeTryToAcquireSpinLockAtDpcLevel() instead of KeAcquireSpinLock();
however this is not sufficient since our KeTryToAcquire...() will BSoD
if the caller already acquired the spinlock.

Note, that *ALL* other instances of the debugger spinlock are acquired
with interrupts disabled, in KdEnterDebugger().